backend/requirements.txt
 # backend/requirements.txt

fastapi
uvicorn[standard]
python-dotenv
httpx
python-multipart
PyMuPDF